requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djongo==1.3.6
dnspython==2.3.0
exceptiongroup==1.1.1
geographiclib==2.0
geopy==2.3.0
googlemaps==4.10.0
h11==0.14.0
idna==3.4
lxml==4.9.2
outcome==1.2.0
pycodestyle==2.10.0
pymongo==3.12.3
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.2.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
